3rd person character update
Did a lot of work to get the 3rd person character controller working with all our new systems. Also adjusted the scale of characters in the environment to better match the 3rd person character. Made changed to the movement script to make it more general and applicable to more use cases, not so specific.